BUNCHO FIXES
So many things
