+++ Bean Man Gets His Cool Back Unity Proj/Assets/Plugins/FMOD/Cache/Editor/FMODStudioCache.asset
@@ -24,7 +24,7 @@ MonoBehaviour:
   - {fileID: 3369702041603077526}
   StringsBanks:
   - {fileID: 6392113596281031596}
-  stringsBankWriteTime: 637426198725470098
+  stringsBankWriteTime: 637426112930359384
   cacheVersion: 5
 --- !u!114 &3369702041603077526
 MonoBehaviour:
@@ -40,7 +40,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   Path: ../Sound Boi/Build/Desktop/Master.bank
   Name: Master
-  lastModified: 637426198725460163
+  lastModified: 637426112930349410
   LoadResult: 0
   FileSizes: []
   Exists: 1
@@ -127,7 +127,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   Path: ../Sound Boi/Build/Desktop/Master.strings.bank
   Name: Master.strings
-  lastModified: 637426198725470098
+  lastModified: 637426112930359384
   LoadResult: 0
   FileSizes:
   - Name:
