+++ Bean Man Gets His Cool Back Unity Proj/Assets/Prefabs/NPC.prefab
@@ -59,6 +59,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children:
   - {fileID: 353742206097845493}
+  - {fileID: 7633175454836699103}
   m_Father: {fileID: 2312038113373758856}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
@@ -124,7 +125,7 @@ GameObject:
   - component: {fileID: 353742206097845493}
   - component: {fileID: 1384248629183027983}
   m_Layer: 0
-  m_Name: Glasses-Sheet_0
+  m_Name: Left Glasses
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -206,6 +207,7 @@ GameObject:
   - component: {fileID: 2312038113373758856}
   - component: {fileID: 2312038113373758857}
   - component: {fileID: 2312038113373758862}
+  - component: {fileID: 8249564989096275865}
   m_Layer: 0
   m_Name: NPC
   m_TagString: NPC
@@ -272,6 +274,7 @@ MonoBehaviour:
   hasSpoken: 0
   glasses: {fileID: 889348222434734441}
   noGlasses: {fileID: 9029286352457924729}
+  actualGlasses: {fileID: 0}
   spriteRenderer: {fileID: 1384248629183027983}
   glassesList:
   - {fileID: 9005065327455587475, guid: 60ed65a7f2515be4c8f137ee62dcf92f, type: 3}
@@ -288,6 +291,108 @@ MonoBehaviour:
   - {fileID: -2820679870303600787, guid: 60ed65a7f2515be4c8f137ee62dcf92f, type: 3}
   isWinner: 0
   gameState: {fileID: 0}
+  player: {fileID: 0}
+  isLeft: 1
+  angle: 90
+  distance: 1
+--- !u!114 &8249564989096275865
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2312038113373758853}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 7103e5a551c323d4c8cc911a9a8abe0c, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  thisNPC: {fileID: 0}
+  leftGlasses: {fileID: 1220874818093408338}
+  rightGlasses: {fileID: 2417638376913200079}
+  BeanMan: {fileID: 0}
+--- !u!1 &2417638376913200079
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 7633175454836699103}
+  - component: {fileID: 1002672527669420592}
+  m_Layer: 0
+  m_Name: Right Glasses
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 0
+--- !u!4 &7633175454836699103
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2417638376913200079}
+  m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
+  m_LocalPosition: {x: 0.01, y: 0, z: -0.25}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_Children: []
+  m_Father: {fileID: 8856990646258134207}
+  m_RootOrder: 1
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!212 &1002672527669420592
+SpriteRenderer:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2417638376913200079}
+  m_Enabled: 1
+  m_CastShadows: 0
+  m_ReceiveShadows: 0
+  m_DynamicOccludee: 1
+  m_MotionVectors: 1
+  m_LightProbeUsage: 1
+  m_ReflectionProbeUsage: 1
+  m_RayTracingMode: 0
+  m_RenderingLayerMask: 1
+  m_RendererPriority: 0
+  m_Materials:
+  - {fileID: 10754, guid: 0000000000000000f000000000000000, type: 0}
+  m_StaticBatchInfo:
+    firstSubMesh: 0
+    subMeshCount: 0
+  m_StaticBatchRoot: {fileID: 0}
+  m_ProbeAnchor: {fileID: 0}
+  m_LightProbeVolumeOverride: {fileID: 0}
+  m_ScaleInLightmap: 1
+  m_ReceiveGI: 1
+  m_PreserveUVs: 0
+  m_IgnoreNormalsForChartDetection: 0
+  m_ImportantGI: 0
+  m_StitchLightmapSeams: 1
+  m_SelectedEditorRenderState: 0
+  m_MinimumChartSize: 4
+  m_AutoUVMaxDistance: 0.5
+  m_AutoUVMaxAngle: 89
+  m_LightmapParameters: {fileID: 0}
+  m_SortingLayerID: 0
+  m_SortingLayer: 0
+  m_SortingOrder: 0
+  m_Sprite: {fileID: 9005065327455587475, guid: 60ed65a7f2515be4c8f137ee62dcf92f,
+    type: 3}
+  m_Color: {r: 1, g: 1, b: 1, a: 1}
+  m_FlipX: 0
+  m_FlipY: 0
+  m_DrawMode: 0
+  m_Size: {x: 0.18, y: 0.08}
+  m_AdaptiveModeThreshold: 0.5
+  m_SpriteTileMode: 0
+  m_WasSpriteAssigned: 1
+  m_MaskInteraction: 0
+  m_SpriteSortPoint: 0
 --- !u!1 &9029286352457924729
 GameObject:
   m_ObjectHideFlags: 0
