+++ Bean Man Gets His Cool Back Unity Proj/Assets/Prefabs/Traffic_Cone.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 8ccc5a90328916047a59159eaabf4b38
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
