+++ Bean Man Gets His Cool Back Unity Proj/Assets/Scripts/EndingsManager.cs
