+++ Bean Man Gets His Cool Back Unity Proj/Assets/Scripts/GlassesFlip.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7103e5a551c323d4c8cc911a9a8abe0c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
