+++ Bean Man Gets His Cool Back Unity Proj/Assets/Sprites/BeanMan_Logo.png.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5017b0abcedc78040abf1f8c0ef2100f
+guid: e25eff135548af742ae887466687aa6e
 TextureImporter:
   internalIDToNameTable: []
   externalObjects: {}
