+++ Bean Man Gets His Cool Back Unity Proj/Assets/Sprites/BirthdayCake-Sheet_0.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d99cdefb6abe50740aae9918ebdade7b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
