+++ Bean Man Gets His Cool Back Unity Proj/Assets/Sprites/BirthdayCake-Sheet_8.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8207fc2ba41b1294281d05cd6c5564b3
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
