+++ Bean Man Gets His Cool Back Unity Proj/Assets/Sprites/BirthdayCakeGlasses.anim
@@ -1,0 +1,97 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!74 &7400000
+AnimationClip:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_Name: BirthdayCakeGlasses
+  serializedVersion: 6
+  m_Legacy: 0
+  m_Compressed: 0
+  m_UseHighQualityCurve: 1
+  m_RotationCurves: []
+  m_CompressedRotationCurves: []
+  m_EulerCurves: []
+  m_PositionCurves: []
+  m_ScaleCurves: []
+  m_FloatCurves: []
+  m_PPtrCurves:
+  - curve:
+    - time: 0
+      value: {fileID: 7501776590650419080, guid: 8966b696bf4f4be419b4c1a3e8dfcab7,
+        type: 3}
+    - time: 0.083333336
+      value: {fileID: 8230374211783763685, guid: 8966b696bf4f4be419b4c1a3e8dfcab7,
+        type: 3}
+    - time: 0.16666667
+      value: {fileID: 4404518139233558079, guid: 8966b696bf4f4be419b4c1a3e8dfcab7,
+        type: 3}
+    - time: 0.25
+      value: {fileID: 9134736028731707729, guid: 8966b696bf4f4be419b4c1a3e8dfcab7,
+        type: 3}
+    - time: 0.33333334
+      value: {fileID: -561677032317078148, guid: 8966b696bf4f4be419b4c1a3e8dfcab7,
+        type: 3}
+    - time: 0.41666666
+      value: {fileID: -1625477451086250944, guid: 8966b696bf4f4be419b4c1a3e8dfcab7,
+        type: 3}
+    - time: 0.5
+      value: {fileID: 104677499578559324, guid: 8966b696bf4f4be419b4c1a3e8dfcab7,
+        type: 3}
+    - time: 0.5833333
+      value: {fileID: 9129564888625375887, guid: 8966b696bf4f4be419b4c1a3e8dfcab7,
+        type: 3}
+    attribute: m_Sprite
+    path: 
+    classID: 212
+    script: {fileID: 0}
+  m_SampleRate: 12
+  m_WrapMode: 0
+  m_Bounds:
+    m_Center: {x: 0, y: 0, z: 0}
+    m_Extent: {x: 0, y: 0, z: 0}
+  m_ClipBindingConstant:
+    genericBindings:
+    - serializedVersion: 2
+      path: 0
+      attribute: 0
+      script: {fileID: 0}
+      typeID: 212
+      customType: 23
+      isPPtrCurve: 1
+    pptrCurveMapping:
+    - {fileID: 7501776590650419080, guid: 8966b696bf4f4be419b4c1a3e8dfcab7, type: 3}
+    - {fileID: 8230374211783763685, guid: 8966b696bf4f4be419b4c1a3e8dfcab7, type: 3}
+    - {fileID: 4404518139233558079, guid: 8966b696bf4f4be419b4c1a3e8dfcab7, type: 3}
+    - {fileID: 9134736028731707729, guid: 8966b696bf4f4be419b4c1a3e8dfcab7, type: 3}
+    - {fileID: -561677032317078148, guid: 8966b696bf4f4be419b4c1a3e8dfcab7, type: 3}
+    - {fileID: -1625477451086250944, guid: 8966b696bf4f4be419b4c1a3e8dfcab7, type: 3}
+    - {fileID: 104677499578559324, guid: 8966b696bf4f4be419b4c1a3e8dfcab7, type: 3}
+    - {fileID: 9129564888625375887, guid: 8966b696bf4f4be419b4c1a3e8dfcab7, type: 3}
+  m_AnimationClipSettings:
+    serializedVersion: 2
+    m_AdditiveReferencePoseClip: {fileID: 0}
+    m_AdditiveReferencePoseTime: 0
+    m_StartTime: 0
+    m_StopTime: 0.6666666
+    m_OrientationOffsetY: 0
+    m_Level: 0
+    m_CycleOffset: 0
+    m_HasAdditiveReferencePose: 0
+    m_LoopTime: 1
+    m_LoopBlend: 0
+    m_LoopBlendOrientation: 0
+    m_LoopBlendPositionY: 0
+    m_LoopBlendPositionXZ: 0
+    m_KeepOriginalOrientation: 0
+    m_KeepOriginalPositionY: 1
+    m_KeepOriginalPositionXZ: 0
+    m_HeightFromFeet: 0
+    m_Mirror: 0
+  m_EditorCurves: []
+  m_EulerEditorCurves: []
+  m_HasGenericRootTransform: 0
+  m_HasMotionFloatCurves: 0
+  m_Events: []
