+++ Bean Man Gets His Cool Back Unity Proj/Assets/Sprites/BirthdayCakeGlasses.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0d77a6d8365c7ab47bcb88389239a1a5
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
