+++ Bean Man Gets His Cool Back Unity Proj/Assets/Sprites/BirthdayCakeNoGlasses.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a622f3a0932966841b6656892d1abaed
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
