+++ Bean Man Gets His Cool Back Unity Proj/Assets/Sprites/Stick.png.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5017b0abcedc78040abf1f8c0ef2100f
+guid: 343f878a8e253c44198d2ee610466a20
 TextureImporter:
   internalIDToNameTable: []
   externalObjects: {}
