+++ Bean Man Gets His Cool Back Unity Proj/Assets/Sprites/Traffic_Cone.png.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5017b0abcedc78040abf1f8c0ef2100f
+guid: 57323cdcb452c6b41a25a812ad7fae07
 TextureImporter:
   internalIDToNameTable: []
   externalObjects: {}
